requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.22
botocore==1.27.22
Django==4.0.6
image==1.5.33
jmespath==1.0.1
Pillow==9.2.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
